+++ src/config/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for hello project.
+WSGI config for learnzhongwen project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.